Rising-Phoenix/SoverignAI/requirements.txt
chainlit
langchain
langchain-google-genai
langchain-community
chromadb
google-generativeai
pypdf
python-docx
python-dotenv